Add AppSet preview to resource node
