+++ ui/src/app/applications/components/resource-details/appset-preview-tab.tsx
@@ -1,4 +1,3 @@
-import {Tabs} from 'argo-ui';
 import * as jsYaml from 'js-yaml';
 import * as monacoEditor from 'monaco-editor';
 import * as React from 'react';
@@ -6,55 +5,11 @@ import {useEffect, useRef, useState} from 'react';
 
 import {MonacoEditor} from '../../../shared/components';
 import * as models from '../../../shared/models';
-import {services} from '../../../shared/services';
-import {ApplicationResourcesDiff} from '../application-resources-diff/application-resources-diff';
+import {AppSetGeneratedAppsDiff} from './appset-generated-apps-diff';
 
 interface AppSetPreviewTabProps {
     appSet: models.ApplicationSet;
 }
-
-function extractErrorMessage(err: any): string {
-    if (err?.response?.body?.message) {
-        return err.response.body.message;
-    }
-    if (err?.response?.body?.error) {
-        return err.response.body.error;
-    }
-    if (err?.message) {
-        return err.message;
-    }
-    return 'Preview failed';
-}
-
-function isPermissionDeniedError(err: any): boolean {
-    if (err?.response?.status === 403) {
-        return true;
-    }
-    const msg = extractErrorMessage(err).toLowerCase();
-    return msg.includes('permission denied') || msg.includes('permissiondenied');
-}
-
-const AppManifestCard = ({app}: {app: models.Application}) => {
-    const yaml = React.useMemo(() => jsYaml.dump(app.spec || {}), [app]);
-    const name = app.metadata?.name || '(unnamed)';
-    const namespace = app.metadata?.namespace || '';
-    return (
-        <div className='white-box' style={{marginTop: '10px'}}>
-            <div className='white-box__details'>
-                <div style={{marginBottom: '8px'}}>
-                    <strong>{namespace ? `${namespace}/${name}` : name}</strong>
-                </div>
-                <MonacoEditor
-                    vScrollBar={false}
-                    editor={{
-                        input: {text: yaml, language: 'yaml'},
-                        options: {readOnly: true, minimap: {enabled: false}, wordWrap: 'on', scrollBeyondLastLine: false}
-                    }}
-                />
-            </div>
-        </div>
-    );
-};
 
 export const AppSetPreviewTab = (props: AppSetPreviewTabProps) => {
     const {appSet} = props;
@@ -62,21 +17,16 @@ export const AppSetPreviewTab = (props: AppSetPreviewTabProps) => {
     const modelRef = useRef<monacoEditor.editor.ITextModel | null>(null);
     const resultRef = useRef<HTMLDivElement | null>(null);
     const [editing, setEditing] = useState(false);
-    const [previewing, setPreviewing] = useState(false);
-    const [error, setError] = useState<string | null>(null);
-    const [permissionDenied, setPermissionDenied] = useState(false);
-    const [generated, setGenerated] = useState<models.Application[] | null>(null);
-    const [currentApps, setCurrentApps] = useState<models.Application[]>([]);
-    const [diffStates, setDiffStates] = useState<models.ResourceDiff[] | null>(null);
-    const [diffWarning, setDiffWarning] = useState<string | null>(null);
+    const [parseError, setParseError] = useState<string | null>(null);
     const [editorKey, setEditorKey] = useState(0);
-    const [resultTab, setResultTab] = useState<string>('diff');
+    const [proposedAppSet, setProposedAppSet] = useState<models.ApplicationSet | null>(null);
+    const [trigger, setTrigger] = useState(0);
 
     useEffect(() => {
-        if ((generated || error) && resultRef.current) {
+        if ((proposedAppSet || parseError) && resultRef.current) {
             resultRef.current.scrollIntoView({behavior: 'smooth', block: 'start'});
         }
-    }, [generated, error]);
+    }, [proposedAppSet, parseError]);
 
     const readEditorYaml = (): string => {
         if (modelRef.current) {
@@ -85,83 +35,23 @@ export const AppSetPreviewTab = (props: AppSetPreviewTabProps) => {
         return initialYaml;
     };
 
-    const onPreview = async () => {
-        setError(null);
-        setPermissionDenied(false);
-        setDiffWarning(null);
+    const onPreview = () => {
+        setParseError(null);
         let parsedSpec: any;
         try {
             parsedSpec = jsYaml.load(readEditorYaml());
         } catch (e: any) {
-            setError(`Invalid YAML: ${e.message || e}`);
+            setParseError(`Invalid YAML: ${e.message || e}`);
+            setProposedAppSet(null);
             return;
         }
         if (!parsedSpec || typeof parsedSpec !== 'object') {
-            setError('Invalid YAML: spec must be an object');
+            setParseError('Invalid YAML: spec must be an object');
+            setProposedAppSet(null);
             return;
         }
-        const parsed: models.ApplicationSet = {...appSet, spec: parsedSpec};
-        setPreviewing(true);
-        try {
-            const [proposedApps, currentAppsResult] = await Promise.all([
-                services.applications.appSetGenerate(parsed),
-                services.applications.appSetGenerate(appSet).catch(e => {
-                    setDiffWarning(`Could not generate current appset output for diff: ${extractErrorMessage(e)}`);
-                    return [] as models.Application[];
-                })
-            ]);
-            setGenerated(proposedApps);
-
-            const currentAppsList = currentAppsResult as models.Application[];
-            setCurrentApps(currentAppsList);
-            const defaultNs = appSet.metadata?.namespace || '';
-            const keyOf = (app: models.Application) => `${app.metadata?.namespace || defaultNs}/${app.metadata?.name}`;
-            const currentByKey = new Map<string, models.Application>();
-            for (const app of currentAppsList) {
-                currentByKey.set(keyOf(app), app);
-            }
-            const proposedByKey = new Map<string, models.Application>();
-            for (const app of proposedApps) {
-                proposedByKey.set(keyOf(app), app);
-            }
-            const allKeys: string[] = [];
-            const seen = new Set<string>();
-            for (const k of [...Array.from(proposedByKey.keys()), ...Array.from(currentByKey.keys())]) {
-                if (!seen.has(k)) {
-                    seen.add(k);
-                    allKeys.push(k);
-                }
-            }
-            const states: models.ResourceDiff[] = allKeys.map(key => {
-                const current = currentByKey.get(key) || null;
-                const proposed = proposedByKey.get(key) || null;
-                const sep = key.indexOf('/');
-                const ns = key.slice(0, sep);
-                const name = key.slice(sep + 1);
-                return {
-                    group: 'argoproj.io',
-                    kind: 'Application',
-                    namespace: ns,
-                    name,
-                    hook: false,
-                    normalizedLiveState: current,
-                    predictedLiveState: proposed,
-                    liveState: current,
-                    targetState: proposed
-                } as models.ResourceDiff;
-            });
-            setDiffStates(states);
-        } catch (e: any) {
-            if (isPermissionDeniedError(e)) {
-                setPermissionDenied(true);
-            }
-            setError(extractErrorMessage(e));
-            setGenerated(null);
-            setDiffStates(null);
-            setCurrentApps([]);
-        } finally {
-            setPreviewing(false);
-        }
+        setProposedAppSet({...appSet, spec: parsedSpec});
+        setTrigger(t => t + 1);
     };
 
     const onEdit = () => {
@@ -189,17 +79,17 @@ export const AppSetPreviewTab = (props: AppSetPreviewTabProps) => {
                         <div>
                             {editing ? (
                                 <>
-                                    <button className='argo-button argo-button--base' onClick={onPreview} disabled={previewing}>
-                                        {previewing ? 'Previewing...' : 'Preview'}
+                                    <button className='argo-button argo-button--base' onClick={onPreview}>
+                                        Preview
                                     </button>{' '}
-                                    <button className='argo-button argo-button--base-o' onClick={onCancelEdit} disabled={previewing}>
+                                    <button className='argo-button argo-button--base-o' onClick={onCancelEdit}>
                                         Cancel
                                     </button>
                                 </>
                             ) : (
                                 <>
-                                    <button className='argo-button argo-button--base' onClick={onPreview} disabled={previewing}>
-                                        {previewing ? 'Previewing...' : 'Preview'}
+                                    <button className='argo-button argo-button--base' onClick={onPreview}>
+                                        Preview
                                     </button>{' '}
                                     <button className='argo-button argo-button--base-o' onClick={onEdit}>
                                         Edit
@@ -224,95 +114,18 @@ export const AppSetPreviewTab = (props: AppSetPreviewTabProps) => {
                 </div>
             </div>
 
-            {error && (
+            {parseError && (
                 <div ref={resultRef} className='white-box' style={{marginTop: '15px'}}>
                     <div className='white-box__details'>
-                        {permissionDenied ? (
-                            <>
-                                <p>PERMISSION DENIED</p>
-                                <div>
-                                    You need <code>applicationsets, create</code> permission on project <code>{project}</code> to preview generated Applications.
-                                </div>
-                            </>
-                        ) : (
-                            <>
-                                <p>PREVIEW FAILED</p>
-                                <pre style={{whiteSpace: 'pre-wrap', wordBreak: 'break-word', maxHeight: '400px', overflow: 'auto'}}>{error}</pre>
-                            </>
-                        )}
+                        <p>PREVIEW FAILED</p>
+                        <pre style={{whiteSpace: 'pre-wrap', wordBreak: 'break-word', maxHeight: '400px', overflow: 'auto'}}>{parseError}</pre>
                     </div>
                 </div>
             )}
 
-            {generated && !error && (
-                <div ref={resultRef} style={{marginTop: '15px'}}>
-                    <div className='white-box'>
-                        <div className='white-box__details'>
-                            <p>GENERATED APPLICATIONS</p>
-                            <div>
-                                {generated.length} application{generated.length === 1 ? '' : 's'} would be generated by this ApplicationSet.
-                            </div>
-                        </div>
-                    </div>
-                    {diffWarning && (
-                        <div className='white-box' style={{marginTop: '15px'}}>
-                            <div className='white-box__details'>{diffWarning}</div>
-                        </div>
-                    )}
-                    <div style={{marginTop: '15px'}}>
-                        <Tabs
-                            key={resultTab}
-                            navTransparent={true}
-                            selectedTabKey={resultTab}
-                            onTabSelected={setResultTab}
-                            tabs={[
-                                {
-                                    title: 'CURRENT',
-                                    key: 'current',
-                                    content:
-                                        currentApps.length === 0 ? (
-                                            <div className='white-box'>
-                                                <div className='white-box__details'>No Applications are currently generated by this ApplicationSet.</div>
-                                            </div>
-                                        ) : (
-                                            <div>
-                                                {currentApps.map(app => (
-                                                    <AppManifestCard key={`${app.metadata?.namespace}/${app.metadata?.name}`} app={app} />
-                                                ))}
-                                            </div>
-                                        )
-                                },
-                                {
-                                    title: 'DIFF',
-                                    key: 'diff',
-                                    content:
-                                        diffStates && diffStates.length > 0 ? (
-                                            <ApplicationResourcesDiff states={diffStates} />
-                                        ) : (
-                                            <div className='white-box'>
-                                                <div className='white-box__details'>No changes — proposed output matches current output.</div>
-                                            </div>
-                                        )
-                                },
-                                {
-                                    title: 'PREVIEW',
-                                    key: 'preview',
-                                    content:
-                                        generated.length === 0 ? (
-                                            <div className='white-box'>
-                                                <div className='white-box__details'>No Applications would be generated by this ApplicationSet.</div>
-                                            </div>
-                                        ) : (
-                                            <div>
-                                                {generated.map(app => (
-                                                    <AppManifestCard key={`${app.metadata?.namespace}/${app.metadata?.name}`} app={app} />
-                                                ))}
-                                            </div>
-                                        )
-                                }
-                            ]}
-                        />
-                    </div>
+            {proposedAppSet && !parseError && (
+                <div ref={resultRef}>
+                    <AppSetGeneratedAppsDiff currentAppSet={appSet} proposedAppSet={proposedAppSet} trigger={trigger} rbacProject={project} />
                 </div>
             )}
         </div>
